fix for Pyrex 0.9.6 (thanks Greg Ewing)
git-svn-id: svn://svn.gna.org/svn/soya/trunk@391 f7f35755-3713-0410-b059-dae289c688fa

+++ soya/c.pxd
@@ -34,7 +34,7 @@ cdef extern from "math.h":
 	float sin (float x)
 	float tan (float x)
 	float exp (float x)
-	float pow (float x, int n)
+	float pow (float x, float n)
 	float ceil (float x)
 	float floor (float x)
